errorx: Move package to selfservice
Because this module is primarily used
in selfservice scenarios, it has been
moved to the selfservice parent.
