+++ driver/registry.go
@@ -15,8 +15,8 @@ import (
 	"github.com/ory/x/dbal"
 
 	"github.com/ory/hive/driver/configuration"
-	"github.com/ory/hive/errorx"
 	"github.com/ory/hive/identity"
+	"github.com/ory/hive/selfservice/errorx"
 	password2 "github.com/ory/hive/selfservice/password"
 	"github.com/ory/hive/session"
 )
