+++ driver/registry_abstract.go
@@ -23,8 +23,8 @@ import (
 	"github.com/ory/herodot"
 
 	"github.com/ory/hive/driver/configuration"
-	"github.com/ory/hive/errorx"
 	"github.com/ory/hive/identity"
+	"github.com/ory/hive/selfservice/errorx"
 	password2 "github.com/ory/hive/selfservice/password"
 	"github.com/ory/hive/session"
 )
