+++ selfservice/errors.go
@@ -13,9 +13,9 @@ import (
 	"github.com/ory/x/stringslice"
 
 	"github.com/ory/hive/driver/configuration"
-	"github.com/ory/hive/errorx"
 	"github.com/ory/hive/identity"
 	"github.com/ory/hive/schema"
+	"github.com/ory/hive/selfservice/errorx"
 	"github.com/ory/hive/x"
 )
