+++ selfservice/errorx/manager_test.go
@@ -17,8 +17,8 @@ import (
 
 	"github.com/ory/herodot"
 
-	. "github.com/ory/hive/errorx"
 	"github.com/ory/hive/internal"
+	. "github.com/ory/hive/selfservice/errorx"
 )
 
 // nolint: staticcheck
