+++ selfservice/handler.go
@@ -10,7 +10,7 @@ import (
 	"github.com/ory/x/urlx"
 
 	"github.com/ory/hive/driver/configuration"
-	"github.com/ory/hive/errorx"
+	"github.com/ory/hive/selfservice/errorx"
 	"github.com/ory/hive/session"
 	"github.com/ory/hive/x"
 )
