+++ selfservice/password/strategy.go
@@ -6,9 +6,9 @@ import (
 	"gopkg.in/go-playground/validator.v9"
 
 	"github.com/ory/hive/driver/configuration"
-	"github.com/ory/hive/errorx"
 	"github.com/ory/hive/identity"
 	"github.com/ory/hive/selfservice"
+	"github.com/ory/hive/selfservice/errorx"
 	"github.com/ory/hive/session"
 	"github.com/ory/hive/x"
 )
